Revert "chore: [IA-778] Move zendesk user identify after ask permission (#3899)"
