+++ ts/features/zendesk/components/__tests__/ZendeskSupportComponent.test.tsx
@@ -7,7 +7,16 @@ import { GlobalState } from "../../../../store/reducers/types";
 import { renderScreenFakeNavRedux } from "../../../../utils/testWrapper";
 import ROUTES from "../../../../navigation/routes";
 import ZendeskSupportComponent from "../ZendeskSupportComponent";
+import {
+  idpSelected,
+  loginSuccess,
+  sessionInformationLoadSuccess
+} from "../../../../store/actions/authentication";
+import { SessionToken } from "../../../../types/SessionToken";
+import { PublicSession } from "../../../../../definitions/backend/PublicSession";
+import { SpidLevelEnum } from "../../../../../definitions/backend/SpidLevel";
 import MockZendesk from "../../../../__mocks__/io-react-native-zendesk";
+import { SpidIdp } from "../../../../../definitions/content/SpidIdp";
 import {
   getZendeskConfig,
   zendeskRequestTicketNumber
@@ -16,6 +25,13 @@ import * as zendeskNavigation from "../../store/actions/navigation";
 import { Zendesk } from "../../../../../definitions/content/Zendesk";
 import { getNetworkError } from "../../../../utils/errors";
 
+const mockPublicSession: PublicSession = {
+  bpdToken: "bpdToken",
+  myPortalToken: "myPortalToken",
+  spidLevel: SpidLevelEnum["https://www.spid.gov.it/SpidL2"],
+  walletToken: "walletToken",
+  zendeskToken: "zendeskToken"
+};
 const mockZendeskConfig: Zendesk = {
   panicMode: false
 };
@@ -36,6 +52,25 @@ describe("the ZendeskSupportComponent", () => {
     const component = renderComponent(store, false);
     store.dispatch(zendeskRequestTicketNumber.success(1));
     expect(component.getByTestId("showTicketsButton")).toBeDefined();
+  });
+  describe("when the user is authenticated with session info", () => {
+    const store = createStore(appReducer, globalState as any);
+    store.dispatch(idpSelected({} as SpidIdp));
+    store.dispatch(
+      loginSuccess({
+        token: "abc1234" as SessionToken,
+        idp: "test"
+      })
+    );
+    describe("and the zendeskToken is defined", () => {
+      store.dispatch(sessionInformationLoadSuccess(mockPublicSession));
+      it("should call setUserIdentity with the zendeskToken", () => {
+        renderComponent(store, false);
+        expect(MockZendesk.setUserIdentity).toBeCalledWith({
+          token: mockPublicSession.zendeskToken
+        });
+      });
+    });
   });
 
   describe("when the user press the zendesk open ticket button", () => {
